chore: refactor and break into modules
it was really hard to navigate through different parts of plugin, so I have broken it down into smaller modules. there are still improvements to be made, but it is a big progress compared to previous state
