+++ commitlint.config.js
@@ -1,1 +1,6 @@
-module.exports = {extends: ['@commitlint/config-conventional']}
+module.exports = {
+  extends: ['@commitlint/config-conventional'],
+  "rules": {
+    "body-max-line-length": [1, "always", 1000],
+  }
+}
