+++ cypress.config.ts
@@ -5,11 +5,9 @@ export default defineConfig({
   e2e: {
     baseUrl: 'http://localhost:3003',
     defaultCommandTimeout: 1000,
-    experimentalSessionAndOrigin: true,
     experimentalRunAllSpecs: true,
     video: process.env.CI ? true : false,
     videoUploadOnPasses: false,
-    screenshotOnRunFailure: process.env.CI ? true : false,
-    excludeSpecPattern: 'cypress/e2e/requestMode.cy.ts'
+    screenshotOnRunFailure: process.env.CI ? true : false
   },
 })
