+++ cypress/e2e/formats.cy.ts
@@ -1,33 +1,37 @@
-it('works with xml', () => {
+describe('response formats', () => {
 
-  cy.api({
-    url: '/xml'
-  })
+  it('works with xml', () => {
 
-});
+    cy.api({
+      url: '/xml'
+    })
 
-it('works with html', () => {
+  });
 
-  cy.api({
-    url: '/html'
-  })
+  it('works with html', () => {
 
-});
+    cy.api({
+      url: '/html'
+    })
 
-it('works with json', () => {
+  });
 
-  cy.api({
-    url: '/json'
-  })
+  it('works with json', () => {
 
-});
+    cy.api({
+      url: '/json'
+    })
 
-it('works with text', () => {
+  });
 
-  cy.api({
-    url: '/text'
-  })
+  it('works with text', () => {
 
-  cy.contains('Hey there 👋').should('be.visible')
+    cy.api({
+      url: '/text'
+    })
+
+    cy.contains('Hey there 👋').should('be.visible')
+
+  });
 
 });
