+++ cypress/e2e/requestMode.cy.ts
@@ -1,10 +1,24 @@
-it('request mode', () => {
+describe('request mode toggle', () => {
 
-  cy.request('DELETE', '/')
-  cy.api('/')
-  cy.request('POST', '/')
+  it('request mode on', { env: { requestMode: true } }, () => {
 
-  cy.get('[data-cy=requestPanel')
-    .should('have.length', 3)
+    cy.request('DELETE', '/')
+    cy.api('/')
+    cy.request('POST', '/')
+
+    cy.get('[data-cy=requestPanel]')
+      .should('have.length', 3)
+
+  });
+
+  it('request mode off', { env: { requestMode: false } }, () => {
+
+    cy.api('/')
+    cy.request('POST', '/')
+
+    cy.get('[data-cy=requestPanel]')
+      .should('have.length', 1)
+
+  });
 
 });
