+++ cypress/e2e/retries.cy.ts
@@ -1,16 +1,20 @@
-// this test will fail on first try, but pass on second
-it('Retry resets api state', { retries: 1 }, () => {
 
-  // @ts-ignore cy.state() has no type definitions
-  const currentRetry = cy.state('runnable')._currentRetry
+describe('retryability', () => {
+  // this test will fail on first try, but pass on second
+  it('Retry resets api state', { retries: 1 }, () => {
 
-  // making sure we don’t see requests from previous retry
-  cy.get('[data-cy="responseBody"]').should('not.exist')
+    // @ts-ignore cy.state() has no type definitions
+    const currentRetry = cy.state('runnable')._currentRetry
 
-  cy.api('/')
+    // making sure we don’t see requests from previous retry
+    cy.get('[data-cy="responseBody"]').should('not.exist')
 
-  cy.get('[data-cy="responseBody"]', { timeout: 50 })
-    .should('be.visible')
-    .and('have.length', currentRetry) // will be 0 on first try, 1 on second
+    cy.api('/')
+
+    cy.get('[data-cy="responseBody"]', { timeout: 50 })
+      .should('be.visible')
+      .and('have.length', currentRetry) // will be 0 on first try, 1 on second
+
+  });
 
 });
