+++ cypress/e2e/session.cy.ts
@@ -1,23 +1,27 @@
-beforeEach(() => {
-  cy.session('login', () => {
-    cy.log('Hello')
+describe('compatibility with sessions', () => {
+
+  beforeEach(() => {
+    cy.session('login', () => {
+      cy.log('Hello')
+    })
+  });
+
+  it('creates a session', () => {
+
+    cy.api('/')
+
+    cy.get('[data-cy=responseBody]')
+      .should('be.visible')
+
   })
+
+  it('uses a session', () => {
+
+    cy.api('/')
+
+    cy.get('[data-cy=responseBody]')
+      .should('be.visible')
+
+  })
+
 });
-
-it('creates a session', () => {
-
-  cy.api('/')
-
-  cy.get('[data-cy=responseBody]')
-    .should('be.visible')
-
-})
-
-it('uses a session', () => {
-
-  cy.api('/')
-
-  cy.get('[data-cy=responseBody]')
-    .should('be.visible')
-
-})
