+++ cypress/e2e/snapshotOnly.cy.ts
@@ -1,31 +1,35 @@
-it('snapshotOnly mode removes plugin UI', {
-  baseUrl: null, env: {
-    snapshotOnly: true
-  }
-}, () => {
+describe('snapshot only mode', () => {
 
-  cy.visit('server-public/test.html')
-  cy.contains('MY PAGE')
-    .should('be.visible')
-  cy.api('http://localhost:3003/')
-  cy.api('http://localhost:3003/')
-  cy.get('[data-cy="responseBody"]')
-    .should('not.exist')
+  it('snapshotOnly mode removes plugin UI', {
+    baseUrl: null, env: {
+      snapshotOnly: true
+    }
+  }, () => {
 
-});
+    cy.visit('server-public/test.html')
+    cy.contains('MY PAGE')
+      .should('be.visible')
+    cy.api('http://localhost:3003/json')
+    cy.api('http://localhost:3003/text')
+    cy.get('[data-cy="responseBody"]')
+      .should('not.exist')
 
-it('snapshotOnly does not affect later plugin use', () => {
+  });
 
-  cy.api('/')
-  cy.get('[data-cy="responseBody"]')
-    .should('be.visible')
+  it('snapshotOnly does not affect later plugin use', () => {
 
-});
+    cy.api('/')
+    cy.get('[data-cy="responseBody"]')
+      .should('be.visible')
 
-it('visiting a page does not affect showing plugin', { baseUrl: null }, () => {
+  });
 
-  cy.api('http://localhost:3003/')
-  cy.visit('server-public/test.html')
-  cy.api('http://localhost:3003/')
+  it('visiting a page does not affect showing plugin', { baseUrl: null }, () => {
+
+    cy.api('http://localhost:3003/')
+    cy.visit('server-public/test.html')
+    cy.api('http://localhost:3003/')
+
+  });
 
 });
