+++ package.json
@@ -44,7 +44,7 @@
     "@typescript-eslint/parser": "^5.42.1",
     "@vitejs/plugin-vue": "^3.1.0",
     "autoprefixer": "^10.4.12",
-    "cypress": "^11.2.0",
+    "cypress": "^12.1.0",
     "eslint": "^8.27.0",
     "eslint-plugin-no-only-tests": "^3.1.0",
     "eslint-plugin-vue": "^9.7.0",
