--- src/utils/getDoc.ts
@@ -1,5 +1,0 @@
-export const getDoc = () => {
-  // @ts-ignore cy.state() has no type definitions
-  const doc: Document = cy.state('document');
-  return doc
-}
